requirements.txt
Pillow
Flask
torch
torchvision
numpy
math
collections
opencv-python
